requirements.txt
requests
Flask
twilio
uWSGI
python-dotenv